feat(ui): unify list shell header and hide meta
