+++ frontend/apps/web/src/pages/ListPage.vue
@@ -1,48 +1,22 @@
 <template>
   <section class="page">
-    <header class="header">
-      <div>
-        <h2>{{ title }}</h2>
-        <p class="meta">{{ subtitle }}</p>
-      </div>
-      <div class="actions">
-        <span class="pill" :class="status">{{ statusLabel }}</span>
-        <button class="ghost" @click="onReload" :disabled="loading">Reload</button>
-      </div>
-    </header>
+    <PageHeader
+      :title="title"
+      :subtitle="subtitle"
+      :status="status"
+      :status-label="statusLabel"
+      :loading="loading"
+      :on-reload="onReload"
+    />
 
-    <section class="toolbar">
-      <div class="search">
-        <input
-          type="search"
-          placeholder="Search"
-          :value="searchTerm"
-          :disabled="loading"
-          @input="onSearchInput"
-          @keydown.enter.prevent="submitSearch"
-        />
-      </div>
-      <div class="filters">
-        <span class="filter">All</span>
-        <span class="filter">Active</span>
-        <span class="filter">Archived</span>
-      </div>
-      <div class="sort">
-        <span class="label">Sort</span>
-        <div class="sort-options">
-          <button
-            v-for="opt in sortOptions"
-            :key="opt.value"
-            type="button"
-            class="sort-option"
-            :class="{ active: opt.value === sortValue }"
-            @click="onSort(opt.value)"
-          >
-            {{ opt.label }}
-          </button>
-        </div>
-      </div>
-    </section>
+    <PageToolbar
+      :loading="loading"
+      :search-term="searchTerm || ''"
+      :sort-options="sortOptions || []"
+      :sort-value="sortValue || ''"
+      :on-search="onSearch"
+      :on-sort="onSort"
+    />
 
     <StatusPanel v-if="loading" title="Loading list..." variant="info" />
     <StatusPanel
@@ -83,8 +57,9 @@
 </template>
 
 <script setup lang="ts">
-import { computed } from 'vue';
 import StatusPanel from '../components/StatusPanel.vue';
+import PageHeader from '../components/page/PageHeader.vue';
+import PageToolbar from '../components/page/PageToolbar.vue';
 import type { StatusError } from '../composables/useStatus';
 
 const props = defineProps<{
@@ -107,18 +82,9 @@ const props = defineProps<{
   searchTerm?: string;
   sortOptions?: Array<{ label: string; value: string }>;
   sortValue?: string;
+  subtitle: string;
+  statusLabel: string;
 }>();
-
-const statusLabel = computed(() => {
-  if (props.status === 'loading') return 'Loading';
-  if (props.status === 'error') return 'Error';
-  if (props.status === 'empty') return 'Empty';
-  return 'Ready';
-});
-
-const subtitle = computed(() => {
-  return `${props.records.length} records · ${props.columns.length} columns`;
-});
 function formatValue(value: unknown) {
   if (Array.isArray(value)) {
     return value.join(', ');
@@ -133,14 +99,6 @@ function handleRow(row: Record<string, unknown>) {
   props.onRowClick(row);
 }
 
-function onSearchInput(event: Event) {
-  const target = event.target as HTMLInputElement;
-  props.onSearch(target.value);
-}
-
-function submitSearch() {
-  props.onSearch(props.searchTerm || '');
-}
 </script>
 
 <style scoped>
@@ -149,88 +107,6 @@ function submitSearch() {
   gap: 16px;
 }
 
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.actions {
-  display: flex;
-  gap: 8px;
-  align-items: center;
-}
-
-.meta {
-  color: #64748b;
-  font-size: 14px;
-}
-
-.toolbar {
-  display: grid;
-  grid-template-columns: 1fr auto auto;
-  gap: 12px;
-  background: white;
-  border-radius: 12px;
-  padding: 12px 16px;
-  box-shadow: 0 14px 24px rgba(15, 23, 42, 0.08);
-  align-items: center;
-}
-
-.search input {
-  width: 100%;
-  padding: 8px 10px;
-  border-radius: 10px;
-  border: 1px solid rgba(15, 23, 42, 0.1);
-  background: #f8fafc;
-}
-
-.filters {
-  display: flex;
-  gap: 8px;
-}
-
-.filter {
-  padding: 6px 10px;
-  border-radius: 999px;
-  background: #f1f5f9;
-  font-size: 12px;
-  color: #475569;
-}
-
-.sort {
-  display: grid;
-  justify-items: end;
-  gap: 4px;
-}
-
-.sort .label {
-  font-size: 11px;
-  text-transform: uppercase;
-  letter-spacing: 0.08em;
-  color: #94a3b8;
-}
-
-.sort-options {
-  display: flex;
-  gap: 6px;
-  flex-wrap: wrap;
-  justify-content: flex-end;
-}
-
-.sort-option {
-  padding: 6px 10px;
-  border-radius: 999px;
-  border: 1px solid rgba(15, 23, 42, 0.12);
-  background: #f8fafc;
-  font-size: 12px;
-  color: #475569;
-}
-
-.sort-option.active {
-  background: #111827;
-  color: #f8fafc;
-}
 
 .table {
   overflow: auto;
@@ -264,53 +140,4 @@ tr:hover {
   cursor: pointer;
 }
 
-.pill {
-  padding: 4px 10px;
-  border-radius: 999px;
-  font-size: 11px;
-  text-transform: uppercase;
-  letter-spacing: 0.08em;
-  background: #e2e8f0;
-  color: #1e293b;
-}
-
-.pill.ok {
-  background: #dcfce7;
-  color: #14532d;
-}
-
-.pill.error {
-  background: #fee2e2;
-  color: #991b1b;
-}
-
-.pill.loading {
-  background: #e0f2fe;
-  color: #075985;
-}
-
-.pill.empty {
-  background: #fef3c7;
-  color: #92400e;
-}
-
-button {
-  padding: 8px 12px;
-  border: none;
-  border-radius: 10px;
-  background: #111827;
-  color: white;
-  cursor: pointer;
-}
-
-.ghost {
-  background: transparent;
-  color: #111827;
-  border: 1px solid rgba(15, 23, 42, 0.12);
-}
-
-button:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-}
 </style>
